renamed LevelManager to LevelSystem
